Deleting arrays package, moving the file to dataStructures package and documenting code
